# docs/api/control-plane-openapi.yaml
openapi: 3.1.0
info:
  title: GuardMCP Control Plane Public API
  version: 1.0.0
  description: >-
    Public API v1 served by services/control-plane for the GuardMCP console.
    Most identifiers and demo data mirror the deterministic seeds in
    infra/postgres/001-demo-seed.sql and infra/redis/001-demo-seed.sh; the
    Replay-only fixtures used to exercise hash-chain and pagination behavior
    (see DemoSeed.SESSION_BROKEN_CHAIN_ID / SESSION_LARGE_ID) are in-memory
    only and intentionally absent from those seed files.
servers:
  - url: http://localhost:8080
    description: Local development (docker compose exposes the same port)
...
paths:
...
  /api/v1/events/recent:
    get:
      tags: [sessions]
      summary: Recent security events (GMCP-80 §3.3)
      description: >-
        Backs the SCR-101 "최근 보안 이벤트" widget and doubles as the SSE
        reconnect gap-fill (4.2 "복구 시 끊긴 구간은 폴링으로 보충").
        `since` accepts either an eventId (resume strictly after it) or an
        ISO-8601 instant (resume from that moment onward); results are
        newest-first. Backed by the real guard_event table (GMCP-24), not the
        Replay demo fixtures.
      operationId: listRecentEvents
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: sessionId
          in: query
          required: false
          schema:
            type: string
        - name: since
          in: query
          required: false
          description: An eventId (uuid) or an ISO-8601 instant
          schema:
            type: string
      responses:
        "200":
          description: Recent events, newest first
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecentEventsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/events/{eventId}:
    get:
      tags: [sessions]
      summary: Single timeline node lookup (deep-link support)
      description: >-
        Same node schema as a timeline entry, with `sessionId` added, so a
        deep link (`/replay/{sessionId}?event={eventId}`) can paint the detail
        panel without loading the full timeline first.
      operationId: getEvent
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: The timeline node, with its owning sessionId
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventLookupResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/events/{eventId}/reveal:
    post:
      tags: [sessions]
      summary: Reveal the unmasked original payload (NFR-04, GMCP-84 §6.3)
      description: >-
        Requires the events:reveal permission (PermissionService): a
        privileged X-Actor-Role (operator/admin) *and* a valid
        X-Operator-Token (server-side secret, configured out of band via
        REVEAL_OPERATOR_TOKEN — unset by default, which disables this
        endpoint entirely). X-Actor-Role alone is a forgeable header with no
        session/auth system behind it anywhere in this service, so the token
        is the actual access control. Every call — granted or denied —
        writes a reveal_audit_log row (result: success /
        denied_no_permission / denied_not_stored). A syntactically invalid
        eventId 404s before the permission check (nothing to log against); a
        well-formed but unknown or not-yet-permitted eventId reaches the
        permission check first, matching the spec's step order. 409 when
        this event's raw payload was never stored (rawPayloadStorageEnabled
        was off at ingest time, or the key configured now cannot decrypt
        what's stored), not 404. Does not match
        apps/console/lib/api/types.ts's RevealContent (masked-line
        reconstruction is out of scope here); this is the spec's own
        response shape.
      operationId: revealEvent
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Actor-Id
          in: header
          required: false
          schema:
            type: string
        - name: X-Actor-Role
          in: header
          required: false
          description: >-
            Must be "operator" or "admin", or the call is rejected with 403
            regardless of X-Operator-Token.
          schema:
            type: string
        - name: X-Operator-Token
          in: header
          required: true
          description: >-
            Must match the server's configured REVEAL_OPERATOR_TOKEN or the
            call is rejected with 403, regardless of X-Actor-Role.
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevealRequest"
      responses:
        "200":
          description: Unmasked payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevealResponse"
        "403":
          description: >-
            Missing the events:reveal permission — X-Operator-Token is
            missing/invalid/unconfigured server-side, or X-Actor-Role is not
            a privileged role. Either way the attempt is still logged
            (result: denied_no_permission).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: No raw payload was stored for this event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  responses:
    BadRequest:
      description: The request is malformed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    NotFound:
      description: The referenced resource does not exist
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable machine-readable reason code
        message:
          type: string
        details:
          type: object
          additionalProperties:
            type: string
...
    Verdict:
      type: string
      description: >-
        Replay/console display vocabulary (PROJECT.md 5.3, SCREEN-SPACE.md 4.3).
        Distinct from GuardAction: mask_then_allow there corresponds to warn
        here, so this API never emits "mask" despite an editorial slip in the
        original replay-api-spec.md task doc.
      enum: [allow, warn, require_approval, block]
    TimelineNodeType:
      type: string
      enum: [USER_INPUT, AGENT_STEP, TOOL_CALL, VERDICT, RESULT]
    ToolCallDirection:
      type: string
      enum: [req, res]
...
    Span:
      type: object
      required: [start, end]
      properties:
        start:
          type: integer
        end:
          type: integer
    Detection:
      type: object
      required: [type, subtype, span, confidence, maskedAs]
      properties:
        type:
          type: string
          description: PII | SECRET | INJ
        subtype:
          type: string
        span:
          $ref: "#/components/schemas/Span"
        confidence:
          type: number
          minimum: 0
          maximum: 1
        maskedAs:
          type: string
          description: Always the masked form; the raw match is never returned by this API.
    VerdictDetail:
      type: object
      required: [matchedPolicyIds, detections, maskDiffRef, hash, prevHash]
      properties:
        matchedPolicyIds:
          type: array
          items:
            type: string
        detections:
          type: array
          items:
            $ref: "#/components/schemas/Detection"
        maskDiffRef:
          type: string
          description: URL the Mask Diff View fetches separately (GET /events/{id}/mask-diff, out of scope)
        hash:
          type: string
        prevHash:
          type: string
          description: Empty string for the first VERDICT node in a session (chain genesis)
    TimelineNode:
      type: object
      required: [eventId, type, ts, summary, detail]
      description: >-
        `detail` is populated only for type=VERDICT; every other node type
        returns it as an explicit null. toolName/direction/argsDigest are only
        present on TOOL_CALL nodes; verdict/riskScore only on VERDICT nodes.
      properties:
        eventId:
          type: string
          format: uuid
        type:
          $ref: "#/components/schemas/TimelineNodeType"
        ts:
          type: string
          format: date-time
        summary:
          type: string
        toolName:
          type: string
        direction:
          $ref: "#/components/schemas/ToolCallDirection"
        argsDigest:
          type: string
        verdict:
          $ref: "#/components/schemas/Verdict"
        riskScore:
          type: integer
          minimum: 0
          maximum: 100
        detail:
          oneOf:
            - $ref: "#/components/schemas/VerdictDetail"
            - type: "null"
        hasRawPayload:
          type: boolean
          description: >-
            GMCP-84 §8.3. Whether this event has a stored, revealable raw
            payload — true only for a VERDICT node whose guard_event row has
            a non-null raw_payload_ref. Absent/false everywhere else.
...
    EventLookupResponse:
      allOf:
        - type: object
          required: [sessionId]
          properties:
            sessionId:
              type: string
              format: uuid
        - $ref: "#/components/schemas/TimelineNode"
...
    SecurityEvent:
      type: object
      required: [id, sessionId, verdict, tool, at]
      description: >-
        apps/console/lib/api/types.ts's SecurityEvent — the SCR-101 widget's
        lean shape, not the full GuardEvent. mask_then_allow collapses into
        warn, matching Verdict's existing 4-value vocabulary.
      properties:
        id:
          type: string
          format: uuid
        sessionId:
          type: string
        verdict:
          $ref: "#/components/schemas/Verdict"
        tool:
          type: string
        target:
          description: No reliable source yet (only argsDigest is stored); always omitted.
          type: string
        at:
          type: string
          format: date-time
    RecentEventsResponse:
      type: object
      required: [events]
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/SecurityEvent"
...
    RevealRequest:
      type: object
      properties:
        reason:
          type: string
          description: Optional; recorded in the audit log entry regardless.
    RevealResponse:
      type: object
      required: [eventId, rawPayload, revealedBy, revealedAt]
      properties:
        eventId:
          type: string
          format: uuid
        rawPayload:
          type: string
          description: Decrypted original payload (RawPayloadStore.decrypt).
        revealedBy:
          type: string
        revealedAt:
          type: string
          format: date-time
...